+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.11.12
 
+* `Repos`:
+  * `Common`:
+    * `JVM`:
+      * Fixes in `ReadFileKeyValueRepo` methods (`values`/`keys`)
+
 ## 0.11.11
 
 * `Crypto`: